Switch options page to svelte (#15)
* feat: switch options page to svelte

* Restore options UI via Snabbdom
