+++ package.json
@@ -57,8 +57,10 @@
 		"rollup": "^2.2.0",
 		"rollup-plugin-css-only": "^2.0.0",
 		"rollup-plugin-node-resolve": "^3.4.0",
-		"rollup-plugin-string": "^3.0.0",
-		"snabbdom": "^0.7.2",
+                "rollup-plugin-string": "^3.0.0",
+                "rollup-plugin-svelte": "^7.1.0",
+                "svelte": "^3.59.1",
+                "snabbdom": "^0.7.2",
 		"tslib": "^1.9.3",
 		"typescript": "^5.4.5",
 		"typescript-eslint": "^8.33.0",
