+++ rollup.config.js
@@ -3,10 +3,12 @@ import css from 'rollup-plugin-css-only';
 import resolve from 'rollup-plugin-node-resolve';
 import commonjs from '@rollup/plugin-commonjs';
 import replace from '@rollup/plugin-replace';
-import { string } from 'rollup-plugin-string'
+import svelte from 'rollup-plugin-svelte';
+import { string } from 'rollup-plugin-string';
 
 const plugins = [
-	resolve(),
+        svelte({ emitCss: true }),
+        resolve(),
 	commonjs(),
 	typescript(),
 	replace({
