+++ src/declarations.d.ts
@@ -1,1 +1,2 @@
 declare module '*.str.css';
+declare module '*.svelte';
